implement basic renaming logic
The position SVGs are now associated with the deck, and not with the
colour on move and that requires a change of the file name.  In the
upgrading process, we will just copy the file to the new location and
rely on the clean-up that happens in the resolution of
https://github.com/gflohr/anki-chess-opening-trainer/issues/12 for
cleaning up after us.